Correcting formulas since julia had trouble handling the addition/subtraction of numbers with large magnitude differences

+++ src/classes.jl
@@ -4,29 +4,29 @@
 
 @with_kw mutable struct Ocean
     volume::BigFloat = 0.0 #only significant for ocean
-    Nfraction::BigFloat = 0.0 #fraction of total mass in resevoir
+    Nfraction::Float64 = 0.0 #fraction of total mass in resevoir
     NMass::BigFloat = 0.0 #mass
 end
 
 @with_kw mutable struct Crust
-    Nfraction::BigFloat = 0.0 #fraction of total mass in resevoir
+    Nfraction::Float64 = 0.0 #fraction of total mass in resevoir
     NMass::BigFloat = 0.0 #mass
 end
 
 @with_kw mutable struct Mantle
-    Nfraction::BigFloat = 0.0 #fraction of total mass in resevoir
+    Nfraction::Float64 = 0.0 #fraction of total mass in resevoir
     NMass::BigFloat = 0.0 #mass
 end
 
 @with_kw mutable struct Atmosphere
-    Nfraction::BigFloat = 0.0 #fraction of Planetary Nitrogen mass in atmosphere
+    Nfraction::Float64 = 0.0 #fraction of Planetary Nitrogen mass in atmosphere
     NMass::BigFloat = 0.0 # mass of Nitrogen in atmosphere
     Mass::BigFloat = 0.0 # mass of atmosphere
 end
 
 @with_kw mutable struct PlanetaryBody
     NMass::BigFloat = 0.0 # Mass of all N on planet
-    NFraction::BigFloat = 1.0
+    NFraction::Float64 = 1.0
     Mass::BigFloat = 0.0 # Mass of Planet
     ocean::Ocean
     crust::Crust
